+++ internal/utils/tfset.go
@@ -1,0 +1,14 @@
+package utils
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+func StringSliceToTerraformSet(values []string) types.Set {
+	elements := []attr.Value{}
+	for _, v := range values {
+		elements = append(elements, types.StringValue(v))
+	}
+	return types.SetValueMust(types.StringType, elements)
+}